Hooking up predefined scene to aruco detection.

+++ ihmc-perception/src/main/java/us/ihmc/perception/objects/DoorSceneObjects.java
@@ -10,7 +10,7 @@ public class DoorSceneObjects
    /** This refers to the edges of the black parts with no margin. The margins included will be wider than this. */
    public static final double DOOR_ARUCO_MARKER_WIDTH = 0.2032;
 
-   public static final long PULL_DOOR_MARKER_ID = 0;
+   public static final int PULL_DOOR_MARKER_ID = 0;
    public static final RigidBodyTransform PULL_DOOR_FRAME_TRANSFORM_TO_MARKER = new RigidBodyTransform(
          new YawPitchRoll(Math.toRadians(180.0), 0.0, Math.toRadians(180.0)),
          new Point3D(0.0, -0.678702 - 0.005 - 0.006, 1.14141 + 0.02)
@@ -19,7 +19,7 @@ public class DoorSceneObjects
          new YawPitchRoll(Math.toRadians(180.0), 0.0, Math.toRadians(180.0)),
          new Point3D(0.0, -0.678702, 1.14141)
    );
-   public static final long PUSH_DOOR_MARKER_ID = 1;
+   public static final int PUSH_DOOR_MARKER_ID = 1;
    public static final RigidBodyTransform PUSH_DOOR_PANEL_TRANSFORM_TO_MARKER = new RigidBodyTransform(
          new YawPitchRoll(0.0, 0.0, 0.0),
          new Point3D(-DoorModelParameters.DOOR_PANEL_THICKNESS / 2.0,
